move spotcallback to spot_utils
